Run the realm-server tests that were never running
Found while validating the CS-12731 weights generator against a real merged
junit report: the report has no suites for several files that exist on disk.
Two mechanisms, both silent. Between them 24 tests were being filtered out of
every shard, and three whole files — 26 more tests — were never loaded at all.

`tests/index.ts` loads from a hand-maintained `ALL_TEST_FILES`. A file missing
from it is never parsed — the sharder still assigns it and still packs a
weight for it, and it reports nothing. Three files were in that state:
`load-links-batching-test.ts`, `server-endpoints/delegate-session-test.ts`,
and `shard-assignment-test.ts`, which arrived with the weighted sharding, so
the invariants it pins were not being checked either. delegate-session also
imports `module` and `test` as named exports, which qunit's CJS entry does not
provide, and names its module with `basename(__filename)`, undefined in an ES
module — it would have thrown on load had it ever been registered. Both
corrected to the form the rest of the suite uses.

The second mechanism is the module filter. Each shard parses every file and
narrows what *runs* with a QUnit name filter built from its assigned files,
`^(?:a-test\.ts|b-test\.ts)(?:\s>\s|:)`, matched against `<module>: <test>`. A
module titled `X | qualifier` is followed by neither separator, so it selects
nothing — indistinguishable from a file with no tests.
`node-realm-test.ts | file stat probing` has never run for that reason.
CS-12731 then added four more while retitling files so the weights generator
could attribute them, which stopped `cpu-profiler-affinity-gate-test.ts` and
`prerender-artifact-sink-test.ts` from running at all. The filter now admits
` | ` as a third separator, which fixes the pre-existing case too.

`ALL_TEST_FILES` and the filter helpers move to `tests/helpers/suite-registry.ts`
so a test can import them. Two tests in `shard-assignment-test.ts` cover the
two failures: every file on disk is registered exactly once and nothing
registered is missing, and the filter selects every top-level module title
declared anywhere in the suite. Both mutation-checked — restoring either bug
fails exactly one of them, naming the file and module.

Verified locally with the pg harness: every qualified module now runs and
passes, `node-realm-test.ts | file stat probing` among them. The two
newly-registered files need the Matrix homeserver, which a local run does not
have; their endpoint and helpers both exist, so CI is the check.

Fixes CS-12759.
