+++ packages/realm-server/tests/helpers/suite-registry.ts
@@ -1,0 +1,242 @@
+// What the suite runs, and how it decides.
+//
+// Two separate mechanisms, easy to confuse. `ALL_TEST_FILES` is the set of
+// files that get parsed at all — a file missing from it is invisible, no
+// matter what CI's sharding says. `buildModuleFilter` then narrows what
+// *runs*, by module name, from the file list a shard was assigned: every
+// shard parses every file and runs only its own modules.
+//
+// Both are checked by tests/shard-assignment-test.ts, because both fail
+// silently. A file left out of the list, or a module whose name the filter
+// cannot express, simply reports no tests — and a suite that runs 29 fewer
+// tests than it did yesterday looks exactly like one that passed.
+
+export const ALL_TEST_FILES: string[] = [
+  './atomic-batch-indexing-test',
+  './atomic-endpoints-test',
+  './auth-client-test',
+  './await-realm-index-settled-test',
+  './billing-test',
+  './card-dependencies-endpoint-test',
+  './card-endpoints-test',
+  './card-html-endpoints-test',
+  './card-source-endpoints-test',
+  './cpu-profiler-affinity-gate-test',
+  './definition-lookup-test',
+  './searchable-parity-diff-test',
+  './migrate-index-to-workspace-test',
+  './file-watcher-events-test',
+  './full-index-on-startup-test',
+  './full-reindex-test',
+  './http-range-test',
+  './range-request-test',
+  './http2-keepalive-test',
+  './indexing-test',
+  './lazy-mount-test',
+  './listener-dispatcher-test',
+  './liveness-verdict-test',
+  './liveness-responder-test',
+  './module-cache-race-test',
+  './module-syntax-test',
+  './network-inflight-tracker-test',
+  './permissions/permission-checker-test',
+  './prerendering-test',
+  './realm-prefix-declaration-test',
+  './prewarm-query-field-test',
+  './prerender-html-split-test',
+  './prerender-html-split-integration-test',
+  './declared-screenshot-persist-test',
+  './declared-screenshots-indexing-test',
+  './prerender-html-reconcile-test',
+  './media-cache-adapter-test',
+  './media-cache-gc-test',
+  './media-cache-serving-test',
+  './media-cache-dsl-test',
+  './prerender-server-test',
+  './prerender-manager-test',
+  './prerender-host-shell-recycle-test',
+  './prerender-artifact-sink-test',
+  './prerender-affinity-activity-test',
+  './prerender-batch-ownership-test',
+  './prerender-cancellation-test',
+  './prerender-never-settles-test',
+  './async-semaphore-test',
+  './page-pool-expansion-test',
+  './page-pool-priority-test',
+  './page-pool-eviction-recovery-test',
+  './page-pool-standby-refill-test',
+  './page-pool-cert-verifier-retry-test',
+  './page-pool-lease-revalidation-test',
+  './page-pool-unresponsive-tab-test',
+  './prerender-deadlock-test',
+  './runtime-exception-capture-test',
+  './clamp-serialized-error-test',
+  './sanitize-for-jsonb-test',
+  './is-json-content-type-test',
+  './bxl-node-smoke-test',
+  './file-size-limit-test',
+  './content-hash-test',
+  './fitted-formats-parity-test',
+  './prerender-diagnostics-persistence-test',
+  './prerender-proxy-test',
+  './prerender-v8-prof-test',
+  './queue-test',
+  './finalize-orphan-reservations-test',
+  './stuck-reservations-test',
+  './mark-failed-job-test',
+  './pg-queue-finalize-test',
+  './finalize-child-fatal-failure-test',
+  './screenshot-card-test',
+  './run-command-task-test',
+  './realm-endpoints-test',
+  './realm-endpoints/dependencies-test',
+  './realm-advisory-locks-test',
+  './realm-cleanup-transaction-test',
+  './data-plane-write-lock-test',
+  './send-event-test',
+  './realm-registry-backfill-test',
+  './realm-registry-reconciler-test',
+  './realm-registry-writes-test',
+  './realm-file-changes-listener-test',
+  './directory-view-refresher-test',
+  './realm-index-updated-listener-test',
+  './jobs-finished-listener-test',
+  './realm-routing-test',
+  './module-cache-invalidation-listener-test',
+  './pg-adapter-connection-error-test',
+  './pg-adapter-subscribe-test',
+  './module-cache-coordination-test',
+  './realm-endpoints/archived-seal-test',
+  './realm-endpoints/readiness-check-test',
+  './realm-endpoints/directory-test',
+  './realm-endpoints/indexing-errors-test',
+  './realm-endpoints/info-test',
+  './realm-endpoints/invalidate-urls-test',
+  './realm-endpoints/lint-test',
+  './realm-endpoints/markdown-test',
+  './realm-endpoints/mtimes-test',
+  './realm-endpoints/permissions-test',
+  './realm-endpoints/cancel-indexing-job-test',
+  './realm-endpoints/publishability-test',
+  './realm-endpoints/reindex-test',
+  './realm-endpoints/screenshot-test',
+  './realm-endpoints/search-test',
+  './realm-endpoints/user-test',
+  './server-endpoints/archive-realm-test',
+  './server-endpoints/authentication-test',
+  './server-endpoints/bot-commands-test',
+  './server-endpoints/bot-registration-test',
+  './server-endpoints/delegate-session-test',
+  './server-endpoints/delete-realm-test',
+  './server-endpoints/download-realm-test',
+  './server-endpoints/federated-types-test',
+  './server-endpoints/index-responses-test',
+  './server-endpoints/maintenance-endpoints-test',
+  './server-endpoints/publish-progress-test',
+  './server-endpoints/queue-status-test',
+  './server-endpoints/realm-lifecycle-test',
+  './server-endpoints/run-command-endpoint-test',
+  './server-endpoints/screenshot-card-endpoint-test',
+  './server-endpoints/search-test',
+  './server-endpoints/skill-validation-test',
+  './serve-index-test',
+  './server-config-test',
+  './server-endpoints/index-counts-test',
+  './server-endpoints/info-test',
+  './server-endpoints/stripe-session-test',
+  './server-endpoints/stripe-webhook-test',
+  './server-endpoints/user-and-catalog-test',
+  './server-endpoints/incoming-webhook-test',
+  './server-endpoints/webhook-commands-test',
+  './server-endpoints/webhook-receiver-test',
+  './server-endpoints/worker-request-test',
+  './transpile-test',
+  './types-endpoint-test',
+  './virtual-network-test',
+  './request-forward-test',
+  './openrouter-passthrough-test',
+  './publish-unpublish-realm-test',
+  './boxel-domain-availability-test',
+  './get-boxel-claimed-domain-test',
+  './claim-boxel-domain-test',
+  './client-telemetry-test',
+  './realm-identifiers-test',
+  './bfm-card-references-test',
+  './package-shim-handler-test',
+  './command-parsing-utils-test',
+  './command-function-name-test',
+  './tool-request-decoding-test',
+  './query-matches-filter-test',
+  './parse-search-url-test',
+  './matches-filter-integration-test',
+  './eq-containment-integration-test',
+  './load-links-batching-test',
+  './search-resource-helpers-test',
+  './superseded-search-surface-removed-test',
+  './search-entry-test',
+  './search-entries-engine-test',
+  './search-bounds-test',
+  './coerce-error-message-test',
+  './retrieve-scoped-css-test',
+  './settled-by-test',
+  './shard-assignment-test',
+  './canonical-url-memo-test',
+  './dependency-normalization-test',
+  './realm-operations-test',
+  './resolve-published-realm-url-test',
+  './fallback-models-test',
+  './host-routing-validation-test',
+  './normalize-realm-meta-value-test',
+  './job-scoped-search-cache-test',
+  './consuming-realm-header-test',
+  './delete-boxel-claimed-domain-test',
+  './realm-auth-test',
+  './queries-test',
+  './remote-prerenderer-test',
+  './markdown-fallback-server-isolation-test',
+  './sanitize-head-html-test',
+  './node-realm-test',
+  './session-room-queries-test',
+  './session-revocation-test',
+  './session-lifetime-test',
+  './indexing-dashboard-test',
+  './indexing-event-sink-test',
+  './skip-query-backed-expansion-test',
+  './worker-request-signature-test',
+  './worker-request-forwarder-test',
+  './worker-reader-test',
+  './worker-job-registration-test',
+  './realm-index-updater-test',
+];
+
+export function parseModules(value: string): string[] {
+  return value
+    .split(/[|,]/)
+    .map((entry) => entry.trim())
+    .filter(Boolean)
+    .map((entry) => entry.replace(/^['"]+|['"]+$/g, ''));
+}
+
+// Matches a test's full name — `<module>: <test>` for a test declared
+// directly in a module, `<module> > <nested> > …: <test>` below that — against
+// the files this shard owns.
+//
+// The trailing group is what makes the match a whole module name rather than a
+// prefix, so `info-test.ts` does not also select `info-test-helpers.ts`. It
+// has to admit every separator that can follow a file name in a module title:
+//
+//   `:`      the test sits directly in the file's module
+//   ` > `    it sits in a module nested inside it
+//   ` | `    the file declares several top-level modules, distinguished by a
+//            qualifier (`node-realm-test.ts | file stat probing`)
+//
+// Omitting the third silently dropped every qualified module from CI.
+export function buildModuleFilter(modulesToMatch: string[]): string {
+  const escaped = modulesToMatch.map((moduleName) => escapeRegex(moduleName));
+  const pattern = `^(?:${escaped.join('|')})(?:\\s>\\s|:|\\s\\|\\s)`;
+  return `/${pattern}/`;
+}
+
+export function escapeRegex(value: string): string {
+  return value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&').replace(/\//g, '\\/');
+}
